[FEAT] : IconInfoCell 구현
