--- orgo/orgo/Screen/Home/MountainDetail/View/IconInfoView.swift
@@ -1,59 +1,0 @@
-//
-//  IconInfoView.swift
-//  orgo
-//
-//  Created by 김태현 on 2023/09/15.
-//
-
-import UIKit
-
-import SnapKit
-import Then
-
-class IconInfoView: BaseView {
-    
-    // MARK: - UI components
-    
-    
-    // MARK: - Variables and Properties
-    
-    
-    // MARK: - Life Cycle
-    
-    override func configureView() {
-        super.configureView()
-        
-        configureInnerView()
-    }
-    
-    override func layoutView() {
-        super.layoutView()
-        
-        configureLayout()
-    }
-    
-    // MARK: - Functions
-    
-}
-
-
-// MARK: - Configure
-
-extension IconInfoView {
-    
-    private func configureInnerView() {
-        
-    }
-    
-}
-
-
-// MARK: - Layout
-
-extension IconInfoView {
-    
-    private func configureLayout() {
-        
-    }
-    
-}
